requirements.txt
# --- Core Framework & API ---
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0

# --- Security & Rate Limiting ---
django-axes==8.3.1
django-ratelimit==4.1.0
django-csp==4.0
django-ipware==7.0.1
django-admin-honeypot-updated-2021==1.2.0
python-ipware==3.0.0

# --- Google API & Auth ---
google-api-python-client==2.192.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
google-api-core==2.30.0
google-auth==2.49.0
googleapis-common-protos==1.73.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
rsa==4.9.1
pyasn1==0.6.2
pyasn1_modules==0.4.2

# --- PDF & OCR Processing ---
pdfplumber==0.11.9
pdf2image==1.17.0
pdfminer.six==20251230
pypdfium2==5.6.0
pytesseract==0.3.13
pillow==12.1.1

# --- Data Science & Machine Learning ---
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0

# --- Email & Communications ---
sendgrid==6.12.5
python-http-client==3.3.7

# --- Utilities & Dependencies ---
python-decouple==3.8
requests==2.32.5
urllib3==2.6.3
charset-normalizer==3.4.5
certifi==2026.2.25
idna==3.11
cryptography==46.0.5
PyJWT==2.11.0
sqlparse==0.5.5
asgiref==3.11.1
tzdata==2025.3
packaging==26.0
MarkupSafe==3.0.3
Werkzeug==3.1.6
cffi==2.0.0
pycparser==3.0
protobuf==6.33.5
proto-plus==1.27.1
httplib2==0.31.2
pyparsing==3.3.2
uritemplate==4.2.0